learn: ts separate to modules, import / export
