+++ index.html
@@ -73,6 +73,6 @@
         <ul id="tasks3"></ul>
       </form>
     </div>
-    <script src="dist/app.js"></script>
+    <script type="module" src="dist/app.js"></script>
   </body>
 </html>
