+++ src/types/types.ts
@@ -1,0 +1,7 @@
+export interface Task {
+  name: string;
+  done: boolean;
+  category?: Category;
+}
+
+export type Category = "general" | "work" | "gym" | "hobby";
