+++ tsconfig.json
@@ -2,6 +2,7 @@
     "compilerOptions": {
         "rootDir": "./src",
         "outDir": "./dist",
-        "target": "ES2015"
+        "target": "ES2015",
+        "module": "ES2015",
     }
 }
